finished task summary panel slide in for clicking on a task item.
