+++ public/stylesheets/index.css
@@ -20,9 +20,9 @@
 
 body {
     position: absolute;
-    width: 100%;
+    width: 100vw;
     height: 100%;
-    overflow-y: hidden;
+    overflow: hidden;
     font-family: Helvetica, Arial, sans-serif;
     font-size: 12px;
     background-color: var(--tertiary-color);
@@ -62,6 +62,7 @@ body {
     height: 100%;
     background-color: #ffffff;
     border-radius: 5px 5px 0 0;
+    z-index: 5;
 }
 
 #summary {
