+++ public/stylesheets/summary.css
@@ -1,26 +1,27 @@
 /* summary child containers */
-.task-recap {
+#task-recap {
     position: absolute;
-    /* width: 100%; */
     height: 100%;
     padding: 20px;
 }
 
-.task-summary {
+#task-details {
     position: absolute;
-    right: 0;
-    width: 100%;
-    height: 100%;
-}
-
-.task-summary-display {
-    position: absolute;
-    background-color: white;
+    left: -100%;
     width: 100%;
     height: 100%;
     padding: 30px;
     border-radius: 5px 5px 0 0;
+}
+
+.task-details-display {
+    /* left: 300px;
+    display: inline;
+    position: absolute;*/
     transition: all 0.4s;
+    z-index: auto;
+    transform: translateX(100%);
+    background-color: white;
 }
 
 /* summary details elements */
@@ -36,14 +37,10 @@
     align-items: center;
 }
 
-#desc-div {
-    padding: 3px;
-}
-
 .summary-inp {
     border: 1px solid white;
     outline: none;
-    padding: 0px;
+    padding: 0px 3px;
     border-radius: 3px;
     font-size: 1.1em;
 }
@@ -59,9 +56,11 @@ textarea.summary-inp {
 }
 
 #summary-deadline,
-#summary-list {
-    padding: 3px;
+#summary-list,
+#summary-desc {
+    padding: 10px 3px;
     flex-basis: 20%;
+    text-align: left;
 }
 
 #summary-list-select {
