+++ views/dashboard-summary.pug
@@ -1,4 +1,4 @@
-div.task-details
-    div.task-recap
+div.task-summary
+    div#task-recap
         p details about the task
-    div.task-summary
+    div#task-details
